+++ src/Core/Admin/BulkEditContentType.php
@@ -8,7 +8,7 @@
 
 namespace Restrictly\Core\Admin;
 
-use Restrictly\Core\Common\RoleHelper;
+use Restrictly\Core\Services\RoleService;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -115,7 +115,7 @@ class BulkEditContentType extends ContentTypeBase {
 							<td>
 								<span class="title restrictly-screen-reader-span-140"><strong><?php esc_html_e( 'Specific Roles:', 'restrictly-access-control' ); ?></strong></span>
 								<?php
-								$available_roles = RoleHelper::get_available_roles();
+								$available_roles = RoleService::get_available_roles();
 
 								/**
 								 * Available roles array used for rendering checkboxes.