+++ src/Core/Admin/StandardEditContentType.php
@@ -9,7 +9,7 @@
 namespace Restrictly\Core\Admin;
 
 use WP_Post;
-use Restrictly\Core\Common\RoleHelper;
+use Restrictly\Core\Services\RoleService;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -121,7 +121,7 @@ class StandardEditContentType extends ContentTypeBase {
 					<td>
 						<span class="title restrictly-screen-reader-span-140"><strong><?php esc_html_e( 'Specific Roles:', 'restrictly-access-control' ); ?></strong></span>
 						<?php
-						$available_roles = RoleHelper::get_available_roles();
+						$available_roles = RoleService::get_available_roles();
 
 						foreach ( $available_roles as $role_key => $role_name ) {
 							printf(